asset_tracker/requirement.txt
# django libraries
Django==4.2.3
djangorestframework==3.14.0
django-seed==0.3.1

# MySQL Library
mysqlclient==2.2.0


























































